+++ octagons/tests/libFuzzer/test_oct33.c
@@ -7,7 +7,7 @@
 
 extern int LLVMFuzzerTestOneInput(const long *data, size_t dataSize) {
 	unsigned int dataIndex = 0;
-	long dim;
+	int dim;
 	FILE *fp;
 	fp = fopen("out33.txt", "w+");
